Marcar consultas

+++ MedicalClinic/src/main/java/com/example/demo/PatientRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 //CRUD refers Create, Read, Update, Delete
 
-public interface PatientRepository extends CrudRepository<Patient, Integer> {
+public interface PatientRepository extends CrudRepository<Patient, Long> {
     Patient findByEmail(String email);
 }
